+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changes
 
 - type of aux provider from `str` to `AuxProvider`
+- prepare reflex update
+- de-duplicate pagination component
+- de-duplicate ingestion roundtrip tests
 
 ### Deprecated